Update Get-InstalledSoftware.ps1
This modified script does the following:
Retrieves Installed Software: It queries the registry for installed software.
Creates Custom Objects: For each software found, it creates a [PSCustomObject] with DisplayName and DisplayVersion properties.
Sorts the Software List: It sorts the list of software objects alphabetically by DisplayName.
Gets FQDN: It retrieves the FQDN of the local computer using [System.Net.Dns]::GetHostEntry($env:computerName).HostName.
Outputs to Console: It displays the sorted list in the console.
Exports to CSV: It saves the sorted list to a CSV file with the FQDN included in the filename.